[cad-enry] refactoring
