+++ list_of_repositories.txt
@@ -1,5 +1,5 @@
 [
-      "https://github.com/navgurukul/newton/",
-      "https://github.com/MicrosoftDocs/pipelines-python-django/",
-      "https://github.com/OmdenaAI/Arabic-Chapter/"
+      "https://github.com/OmdenaAI/Arabic-Chapter/",
+      "https://github.com/Azure-Samples/resource-manager-python-template-deployment/",
+      "https://github.com/MicrosoftDocs/pipelines-python-django/"
 ]
